Renaming the package name and making changes to path respectively
