+++ exporters/aws-otel-otlp-udp-exporter/src/main/java/software/amazon/opentelemetry/exporters/otlp/udp/trace/OtlpUdpSpanExporterBuilder.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.opentelemetry.exporters.otlp.udp;
+package software.amazon.opentelemetry.exporters.otlp.udp.trace;
 
 import static java.util.Objects.requireNonNull;
